move JobsTable component
